+++ controllers/contacts.js
@@ -1,12 +1,5 @@
-const Joi = require("joi");
 const { HttpError, ctrlWrapper } = require("../helpers");
 const contacts = require("../models/contacts");
-
-const addSchema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string().email().required(),
-  phone: Joi.string().required(),
-});
 
 const getAll = async (req, res, next) => {
   const contactsList = await contacts.listContacts();
@@ -22,19 +15,11 @@ const getById = async (req, res, next) => {
 };
 
 const add = async (req, res, next) => {
-  const { error } = addSchema.validate(req.body);
-  if (error) {
-    throw HttpError(400, error.message);
-  }
   const contact = await contacts.addContact(req.body);
   res.status(201).json(contact);
 };
 
 const updateById = async (req, res, next) => {
-  const { error } = addSchema.validate(req.body);
-  if (error) {
-    throw HttpError(400, error.message);
-  }
   const contact = await contacts.updateContact(req.params.contactId, req.body);
   res.status(200).json(contact);
 };
